requirements.txt
Django~=3.1
python-dateutil=2.8.1
